+++ controllers/usersController.js
@@ -1,4 +1,5 @@
 const User = require("../models/user");
+const { check, validationResult } = require('express-validator');
 
 module.exports = {
   showLoginPage: function(req, res) {
@@ -9,5 +10,20 @@ module.exports = {
   },
   showDashboardPage: function(req, res) {
     res.render('dashboard')
+  },
+  saveUser: function(req, res) {
+    const errors = validationResult(req);
+    if(!errors.isEmpty()) {
+      console.log(errors);
+    } else {
+      let newUser = {
+        name: req.body.name,
+        email: req.body.email,
+        password: req.body.password,
+        passwordConfirmation: req.body.passwordConfirmation
+      }
+      console.log(newUser);
+      res.render('dashboard', {userData: req.body})
+    }
   }
 };
