+++ package.json
@@ -21,8 +21,6 @@
     "express": "^4.17.1",
     "express-ejs-layouts": "^2.5.0",
     "express-validator": "^6.10.0",
-    "http-status-codes": "^2.1.4",
-    "mongodb": "^3.6.4",
-    "mongoose": "^5.11.19"
+    "http-status-codes": "^2.1.4"
   }
 }
